+++ Diagnostics/CMakeLists.txt
@@ -3,14 +3,14 @@
 # )
 
 
-add_library(CorePlugins STATIC
-    ProcessMCMCPlugin.cpp
-    DiagMCMCPlugin.cpp
-    GetPenaltyTermPlugin.cpp
-)
+# add_library(CorePlugins STATIC
+#     ProcessMCMCPlugin.cpp
+#     DiagMCMCPlugin.cpp
+#     GetPenaltyTermPlugin.cpp
+# )
 
-target_link_libraries(CorePlugins PRIVATE MaCh3::All)
-target_include_directories(CorePlugins PUBLIC $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/include>)
+# target_link_libraries(CorePlugins PRIVATE MaCh3::All)
+# target_include_directories(CorePlugins PUBLIC $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/include>)
 
 # set_target_properties(Plugins PROPERTIES
 #     PUBLIC_HEADER "${HEADERS}"
@@ -22,7 +22,15 @@ target_include_directories(CorePlugins PUBLIC $<BUILD_INTERFACE:${CMAKE_BINARY_D
 #         LIBRARY DESTINATION lib/
 #         PUBLIC_HEADER DESTINATION include/Plugins)
 
-add_library(MaCh3::CorePlugins ALIAS CorePlugins)
+# add_library(MaCh3::CorePlugins ALIAS CorePlugins)
+
+# set(CLI_MODULES
+#     ProcessMCMCPlugin.cpp
+#     DiagMCMCPlugin.cpp
+#     GetPenaltyTermPlugin.cpp
+# )
+
+# list(TRANSFORM CLI_MODULES PREPEND "../cli/modules/")
 
 add_custom_target(DiagApps)
 
@@ -40,7 +48,7 @@ foreach(app
     )
     add_executable(${app} ${app}.cpp)
     add_dependencies(DiagApps ${app})
-    target_link_libraries(${app} MaCh3::CorePlugins MaCh3::All MaCh3Warnings ROOT::ROOTDataFrame)  # dont need plugins in MaCh3::All at root scope as only used here. Once all transferred to plugins wont need MaCh3::All either
+    target_link_libraries(${app} MaCh3::CoreBinaryModules MaCh3::All MaCh3Warnings ROOT::ROOTDataFrame)  # dont need plugins in MaCh3::All at root scope as only used here. Once all transferred to plugins wont need MaCh3::All either
 
     install(TARGETS ${app} DESTINATION bin)
 endforeach(app)
