+++ cli/CMakeLists.txt
@@ -1,4 +1,20 @@
-    add_executable(mach3 mach3.cpp MaCh3Program.cpp)
-    target_link_libraries(mach3 PRIVATE MaCh3::CorePlugins MaCh3::All)
+add_subdirectory(modules)
+# set(CLI_MODULES
+#     ProcessMCMCPlugin.cpp
+#     DiagMCMCPlugin.cpp
+#     GetPenaltyTermPlugin.cpp
+# )
 
-    install(TARGETS mach3 DESTINATION bin)
+# list(TRANSFORM CLI_MODULES PREPEND "modules/")
+
+
+
+add_executable(mach3
+    mach3.cpp
+    MaCh3Program.cpp
+    # ${CLI_MODULES}
+)
+
+target_link_libraries(mach3 PRIVATE MaCh3::CoreBinaryModules MaCh3::All)
+
+install(TARGETS mach3 DESTINATION bin)
