+++ cli/modules/DiagMCMCPlugin.hpp
